moved all to lib/commands
